requirements.txt
pdfminer.six
anthropic
openai
python-dotenv
